Refactor PushUseCase and PushUseCaseStrategy.
