+++ app/src/myanmar/java/org/eyeseetea/malariacare/domain/usecase/strategies/PushUseCaseStrategy.java
@@ -1,13 +1,20 @@
 package org.eyeseetea.malariacare.domain.usecase.strategies;
 
-import org.eyeseetea.malariacare.domain.usecase.push.PushUseCase;
+import org.eyeseetea.malariacare.domain.boundary.IPushController;
+import org.eyeseetea.malariacare.domain.boundary.executors.IMainExecutor;
+import org.eyeseetea.malariacare.domain.boundary.repositories.IOrganisationUnitRepository;
+import org.eyeseetea.malariacare.domain.boundary.repositories.ISurveyRepository;
+import org.eyeseetea.malariacare.domain.usecase.push.SurveysThresholds;
 import org.eyeseetea.malariacare.domain.usecase.push.strategies.APushUseCaseStrategy;
 
 public class PushUseCaseStrategy extends APushUseCaseStrategy {
-    public PushUseCaseStrategy(PushUseCase pushUseCase,
-            IPushController pushController,
-            IOrganisationUnitRepository organisationUnitRepository) {
-        super(pushUseCase, pushController, organisationUnitRepository);
+    public PushUseCaseStrategy(IPushController pushController,
+            IOrganisationUnitRepository organisationUnitRepository,
+            SurveysThresholds surveysThresholds,
+            ISurveyRepository surveyRepository,
+            IMainExecutor mainExecutor) {
+        super(pushController, organisationUnitRepository, surveysThresholds, surveyRepository,
+                mainExecutor);
     }
